+++ web/static/js/index.js
@@ -1,0 +1,8 @@
+var app = angular.module("Plex-Sync", []);
+
+app.controller("indexImages", function($scope, $http){
+    $http.get("/api/shows")
+    .then(function(response) {
+        $scope.records = response.data;
+    });
+})