+++ CHANGELOG.md
@@ -11,10 +11,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+* New - Create API for web interface calls.
 * Changed - ([#70](https://github.com/danstis/rmstale/issues/70)) Move settings into Database.
 * Changed - Refactor logging to new logger package.
 * Changed - ([#57](https://github.com/danstis/rmstale/issues/57)) Update web interface settings page to enable population of all settings.
-* New - Create API for web interface calls.
+* Changed - Index images retreved from API.
 
 ## [v0.5.2]